Rename everything to brainmapper and explicitly support old cellfinder directories
