+++ tests/data/cellfinder_output/registration/brainreg.json
@@ -1,0 +1,1 @@
+{"image_paths": "/img.tiff", "backend": "niftyreg", "voxel_sizes": ["10.0", "10.0", "10.0"], "orientation": "asr", "data_orientation": "asr", "brain_geometry": "geometry_keys.Full brain", "save_original_orientation": "False", "atlas": "allen_mouse_100um", "atlas_key": "atlas_key.allen_mouse_50um", "registration_output_folder": "/brainreg_output", "affine_n_steps": "6", "affine_use_n_steps": "5", "freeform_n_steps": "6", "freeform_use_n_steps": "4", "bending_energy_weight": "0.95", "grid_spacing": "10", "smoothing_sigma_reference": "1", "smoothing_sigma_floating": "1.0", "histogram_n_bins_floating": "128.0", "histogram_n_bins_reference": "128.0", "debug": "False"}
