+++ tests/tests/test_unit/test_brainreg_utils.py
@@ -5,7 +5,7 @@ from brainglobe_napari_io.brainreg import utils
 brainreg_dir = (
     pathlib.Path(__file__).parent.parent.parent
     / "data"
-    / "wholebrain_cell_output"
+    / "brainmapper_output"
     / "registration"
 )
